requirements.txt
beautifulsoup4
flake8
numba
numpy
pandas
python-dotenv
requests
selenium
